User commit

+++ src/main/java/com/example/demo/service/StudentServiceimple/StudentServiceImplementation.java
@@ -24,7 +24,7 @@ public Optional<Student> fetchDataById(Long id){
     return repo.findById(id);
 }
 @Override
-public void deleteData(int id){
+public void deleteData(Long id){
     repo.deleteById(id);
 }
 }
